pyproject.toml
[tool.poetry]
name = "commitgen"
version = "1.0.0"
description = "AI Git Commit Message Generator (Powered by Relax AI)"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
python-dotenv = "^1.0.0"
requests = "^2.28.0"

[[tool.poetry.packages]]
include = "cli.py"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
commitgen = "cli:main"